vimpager: expose vim's --startuptime FILE option
Adds a command line option `--startuptime FILE`.

This option simply defers to vim's builtin option of the same name and
format, which writes startup times to the given file. This is extremely
useful for profiling, and given that vimpager is a replacement for other
fast pagers like `more` or `less`, it's really useful for users to be
able to profile it easily.